dcwj_sys/config.py
"""
配置文件
定义应用的配置类
"""
import os
from dotenv import load_dotenv

# 加载环境变量
load_dotenv()


class Config:
    """基础配置类"""
    # 密钥配置
    SECRET_KEY = os.getenv('SECRET_KEY', 'dev-secret-key-please-change')

    # 数据库配置
    SQLALCHEMY_DATABASE_URI = os.getenv('DATABASE_URL', 'mysql+pymysql://root:admin123@localhost:3306/dcwj')
    SQLALCHEMY_TRACK_MODIFICATIONS = False
    SQLALCHEMY_ECHO = False

    # 会话配置
    PERMANENT_SESSION_LIFETIME = 3600 * 24 * 7  # 7天

    # 分页配置
    SURVEYS_PER_PAGE = 10

    # 上传配置
    MAX_CONTENT_LENGTH = 16 * 1024 * 1024  # 16MB


class DevelopmentConfig(Config):
    """开发环境配置"""
    DEBUG = True
    SQLALCHEMY_ECHO = False


class ProductionConfig(Config):
    """生产环境配置"""
    DEBUG = False
    SQLALCHEMY_ECHO = False


# 配置字典
config = {
    'development': DevelopmentConfig,
    'production': ProductionConfig,
    'default': DevelopmentConfig
}
