dcwj_sys/requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.0
SQLAlchemy==2.0.0
PyMySQL==1.1.0
Jinja2==3.1.2
Werkzeug==2.2.2
python-dotenv==1.0.0
